Cleaned classes - added methods to classes - clened method codes. cleaning cleany
